+++ CHANGES.md
@@ -1,5 +1,9 @@
 # CHANGES
-This file contains the log of changes of **RaspberryPI** eXPerience.
+This file contains the log of changes of **RaspberryPi** eXPerience.
+
+
+## 0.5.0 2022-04-09 New eXPerience release
+- Added Serial bi-directional BUS to parallel input/output experience
 
 
 ## 0.4.0 2022-04-01 New eXPerience release
@@ -7,11 +11,11 @@ This file contains the log of changes of **RaspberryPI** eXPerience.
 
 
 ## 0.3.0 2022-03-20 New eXPerience release
-- Added Parallel-to-serial input through shift registers experience
+- Added Parallel-to-serial input experience
 
 
 ## 0.2.0 2022-03-15 New eXPerience release
-- Added Serial-to-parallel output through shift registers experience
+- Added Serial-to-parallel output experience
 
 
 ## 0.1.0 2022-03-15 New eXPerience release